lab-02 p01 - p09 finished

+++ URI_Midterm/p-1008/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
 
         int n = inp.nextInt();
         int whn = inp.nextInt();
-        double aph = inp.nextDouble();
+        float aph = inp.nextFloat();
 
         double s = whn*aph;
